Move GUI files to package
