+++ src/main/java/graphicaluserinterface/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package graphicaluserinterface;
 
 import java.io.IOException;
 import java.util.Collections;
