+++ src/main/java/graphicaluserinterface/MainWindow.java
@@ -1,7 +1,7 @@
-package duke;
+package graphicaluserinterface;
 
+import duke.Duke;
 import duke.ui.Ui;
-
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -9,7 +9,6 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
  */
